+++ shortcuts/actions/conditions.py
@@ -1,12 +1,18 @@
 from shortcuts.actions.base import BaseAction, ChoiceField, Field, GroupIDField
 
 
+IF_CHOICES = (
+    'Equals',
+    'Contains',
+)
+
+
 class IfAction(BaseAction):
-    '''If: you must specify group_id of this if condition'''
+    '''If'''
     itype = 'is.workflow.actions.conditional'
     keyword = 'if'
 
-    condition = ChoiceField('WFCondition', choices=('Equals', 'Contains'), capitalize=True)
+    condition = ChoiceField('WFCondition', choices=IF_CHOICES, capitalize=True)
     compare_with = Field('WFConditionalActionString')
     group_id = GroupIDField('GroupingIdentifier')
 
@@ -16,7 +22,7 @@ class IfAction(BaseAction):
 
 
 class ElseAction(BaseAction):
-    '''Else: else for a specified group_id'''
+    '''Else'''
     itype = 'is.workflow.actions.conditional'
     keyword = 'else'
 
@@ -28,7 +34,7 @@ class ElseAction(BaseAction):
 
 
 class EndIfAction(BaseAction):
-    '''EndIf: end a condition with specified group_id'''
+    '''EndIf: end a condition'''
     itype = 'is.workflow.actions.conditional'
     keyword = 'endif'