+++ shortcuts/loader.py
@@ -1,7 +1,7 @@
 import collections
 import copy
 import plistlib
-from typing import TYPE_CHECKING, Any, Dict, List, TextIO, Type, Union
+from typing import TYPE_CHECKING, Any, Dict, List, TextIO, Tuple, Type, Union
 
 import toml
 
@@ -90,7 +90,7 @@ class PListLoader(BaseLoader):
             f.name: f._attr for f in action_class().fields
         }
         params = {
-            shortcut_name_to_field_name[p]: cls._load_parameter_value(v)
+            shortcut_name_to_field_name[p]: WFDeserializer(v).deserialized_data
             for p, v in action_dict['WFWorkflowActionParameters'].items()
             if p in shortcut_name_to_field_name
         }
@@ -102,67 +102,161 @@ class PListLoader(BaseLoader):
         identifier = action_dict['WFWorkflowActionIdentifier']
         action_class = ITYPE_TO_ACTION_MAP.get(identifier)
 
+        # in some cases we have action classes with the same itype (identifier)
+        # but they have different parameters
+        #
+        # For example, Base64EncodeAction and Base64DecodeAction
+        # classes have the same itype: "is.workflow.actions.base64encode"
+        # we need to check parameters to determine which action class we need to use
+        #
         # todo: action-based common solution
         if identifier == 'is.workflow.actions.conditional':
-            from shortcuts.actions import IfAction, ElseAction, EndIfAction
-            flow_to_action = {a.default_fields['WFControlFlowMode']: a for a in (IfAction, ElseAction, EndIfAction)}  # type: ignore
-            action_params = action_dict['WFWorkflowActionParameters']
-            action_class = flow_to_action[action_params['WFControlFlowMode']]
-
-        if identifier == 'is.workflow.actions.base64encode':
-            from shortcuts.actions import Base64EncodeAction, Base64DecodeAction
-            action_params = action_dict['WFWorkflowActionParameters']
-            if action_params.get('WFEncodeMode') == 'Encode':
-                action_class = Base64EncodeAction
-            elif action_params.get('WFEncodeMode') == 'Decode':
-                action_class = Base64DecodeAction
-
-        if identifier == 'is.workflow.actions.repeat.count':
-            from shortcuts.actions import RepeatEndAction, RepeatStartAction
-            flow_to_action = {a.default_fields['WFControlFlowMode']: a for a in (RepeatEndAction, RepeatStartAction)}  # type: ignore
-            action_params = action_dict['WFWorkflowActionParameters']
-            action_class = flow_to_action[action_params['WFControlFlowMode']]
+            action_class = cls._get_if_else_action_class(action_dict)
+        elif identifier == 'is.workflow.actions.base64encode':
+            action_class = cls._get_base64_action_class(action_dict)
+        elif identifier == 'is.workflow.actions.repeat.count':
+            action_class = cls._get_repeat_action_class(action_dict)
+        elif identifier == 'is.workflow.actions.choosefrommenu':
+            action_class = cls._get_menu_action_class(action_dict)
 
         return action_class
 
     @classmethod
-    def _load_parameter_value(cls, value: Union[Dict, str]) -> Union[str, List, Dict]:
-        # todo: move to fields
-        if not isinstance(value, dict):
-            return value
-
-        serialization_type = value.get('WFSerializationType')
-
-        # todo: refactor
-        if serialization_type == 'WFTextTokenString':
-            return cls._get_variable_string(value)
-        elif serialization_type == 'WFDictionaryFieldValue':
-            return cls._load_dictionary(value)
-        elif serialization_type == 'WFTextTokenAttachment':
-            return cls._load_text_token_attachment(value)
-        elif serialization_type == 'WFTokenAttachmentParameterState':
-            return cls._load_parameter_value(value['Value'])
-        else:
-            raise RuntimeError(f'Unknown parameter serialization type: {value.get("WFSerializationType")}')
+    def _get_if_else_action_class(cls, action_dict: Dict) -> Type['BaseAction']:
+        """Returns If-Else-EndIf action class based on WFControlFlowMode parameter of action_dict"""
+        from shortcuts.actions import IfAction, ElseAction, EndIfAction
+        return cls._get_action_class_by_wf_control_flow(
+            from_classes=(IfAction, ElseAction, EndIfAction),
+            action_dict=action_dict,
+        )
 
     @classmethod
-    def _load_dictionary(cls, variable_dict: Dict) -> List[Dict[str, Any]]:
+    def _get_repeat_action_class(cls, action_dict: Dict) -> Type['BaseAction']:
+        """Returns Repeat-Start or Repeat-End action class based on WFControlFlowMode parameter of action_dict"""
+        from shortcuts.actions import RepeatEndAction, RepeatStartAction
+        return cls._get_action_class_by_wf_control_flow(
+            from_classes=(RepeatEndAction, RepeatStartAction),
+            action_dict=action_dict,
+        )
+
+    @classmethod
+    def _get_menu_action_class(cls, action_dict: Dict) -> Type['BaseAction']:
+        """Returns Menu action class based on WFControlFlowMode parameter of action_dict"""
+        from shortcuts.actions import MenuStartAction, MenuItemAction, MenuEndAction
+        return cls._get_action_class_by_wf_control_flow(
+            from_classes=(MenuStartAction, MenuItemAction, MenuEndAction),
+            action_dict=action_dict,
+        )
+
+    @classmethod
+    def _get_action_class_by_wf_control_flow(cls,
+                                             from_classes: Tuple[Type['BaseAction'], ...],
+                                             action_dict: Dict) -> Type['BaseAction']:
+        """Returns class from `from_classes` based on WFControlFlowMode parameter of action_dict"""
+        flow_to_action = {a.default_fields['WFControlFlowMode']: a for a in from_classes}  # type: ignore
+        wf_control_flow_mode = action_dict['WFWorkflowActionParameters']['WFControlFlowMode']
+        return flow_to_action[wf_control_flow_mode]
+
+    @classmethod
+    def _get_base64_action_class(cls, action_dict: Dict) -> Type['BaseAction']:
+        """Returns Base64EncodeAction or Base64DecodeAction based on WFEncodeMode parameter of action_dict"""
+        from shortcuts.actions import Base64EncodeAction, Base64DecodeAction
+        action_params = action_dict['WFWorkflowActionParameters']
+
+        # by default it's encode, even if it doesn't have WFEncodeMode parameter
+        encode_mode = action_params.get('WFEncodeMode', 'Encode')
+        if encode_mode == 'Encode':
+            return Base64EncodeAction
+        elif encode_mode == 'Decode':
+            return Base64DecodeAction
+
+        raise RuntimeError(f'Unknown WFEncodeMode: "{encode_mode}"')
+
+
+class WFDeserializer:
+    """
+    Deserializer for WF fields (from shortcuts plist)
+    which converts their data to a format acceptable by Actions
+    """
+    def __init__(self, data) -> None:
+        self._data = data
+
+    @property
+    def deserialized_data(self) -> Union[str, List, Dict]:
+        if not isinstance(self._data, dict):
+            # todo: check if there are other types
+            return self._data
+
+        serialization_to_field_map: Dict[str, Type[WFDeserializer]] = {
+            'WFTextTokenString': WFVariableStringField,
+            'WFDictionaryFieldValue': WFDictionaryField,
+            'WFTextTokenAttachment': WFTextTokenAttachmentField,
+            'WFTokenAttachmentParameterState': WFTokenAttachmentParameterStateField,
+        }
+
+        deserializer = serialization_to_field_map[self._data.get('WFSerializationType')]  # type: ignore
+        if deserializer:
+            return deserializer(self._data).deserialized_data
+
+        raise RuntimeError(f'Unknown serialization type: {self._data.get("WFSerializationType")}')
+
+
+class WFTokenAttachmentParameterStateField(WFDeserializer):
+    def __init__(self, data) -> None:
+        self._data = data['Value']
+
+
+class WFTextTokenAttachmentField(WFDeserializer):
+    @property
+    def deserialized_data(self) -> str:
+        if self._data['Value'].get('Type') == 'Ask':
+            return '{{ask_when_run}}'
+
+        return self._data['Value']['VariableName']
+
+
+class WFDictionaryField(WFDeserializer):
+    @property
+    def deserialized_data(self) -> List[Dict[str, Any]]:
         result = []
-        for item in variable_dict['Value']['WFDictionaryFieldValueItems']:
-            key = cls._load_parameter_value(item['WFKey'])
-            value = cls._load_parameter_value(item['WFValue'])
+        for item in self._data['Value']['WFDictionaryFieldValueItems']:
+            key = WFDeserializer(item['WFKey']).deserialized_data
+            value = WFDeserializer(item['WFValue']).deserialized_data
             result.append({'key': key, 'value': value})
         return result
 
-    @classmethod
-    def _load_text_token_attachment(cls, variable_dict: Dict):
-        return variable_dict['Value']['VariableName']
 
-    @classmethod
-    def _get_variable_string(cls, variable_dict: Dict) -> str:
+class WFVariableStringField(WFDeserializer):
+    """
+    Converts wf variable string (dictionary)
+        <dict>
+            <key>Value</key>
+            <dict>
+                <key>attachmentsByRange</key>
+                <dict>
+                    <key>{7, 1}</key>
+                    <dict>
+                        <key>Type</key>
+                        <string>Variable</string>
+                        <key>VariableName</key>
+                        <string>name</string>
+                    </dict>
+                </dict>
+                <key>string</key>
+                <string>Hello, ￼!</string>
+            </dict>
+            <key>WFSerializationType</key>
+            <string>WFTextTokenString</string>
+        </dict>
+
+    to a shortcuts-string:
+        "Hello, {{var}}!"
+    """
+    @property
+    def deserialized_data(self) -> str:
         # if this field is a string with variables,
         # we need to convert it to our representation
-        value = variable_dict['Value']
+        value = self._data['Value']
         value_string = value['string']
 
         positions = {}
@@ -175,7 +269,7 @@ class PListLoader(BaseLoader):
                 )
 
             # let's find positions of all variables in the string
-            position = cls._get_position(variable_range)
+            position = self._get_position(variable_range)
             positions[position] = '{{%s}}' % variable_data['VariableName']
 
         # and then replace them with '{{variable_name}}'
@@ -186,7 +280,6 @@ class PListLoader(BaseLoader):
 
         return value_string
 
-    @classmethod
-    def _get_position(cls, range_str: str) -> int:
+    def _get_position(self, range_str: str) -> int:
         ranges = list(map(lambda x: int(x.strip()), range_str.strip('{} ').split(',')))
         return ranges[0]